+++ src/components/HeaderExpanded/HeaderExpanded.module.scss
@@ -10,6 +10,7 @@
     -webkit-box-shadow: 0px 15px 30px #262B3414;
     -moz-box-shadow: 0px 15px 30px #262B3414;
          box-shadow: 0px 15px 30px #262B3414;
+         
     .header{
         display: flex;
         align-items: flex-end;
