+++ src/pages/ProfilePage/ProfilePage.module.scss
@@ -86,5 +86,4 @@
             }
         }
     }
-
 }
