--- src/pages/ProfilePage/ProfileSubPage/AboutUs/AboutUs.js
@@ -1,14 +1,0 @@
-import React, { Component } from "react";
-import { Grid } from "@material-ui/core";
-
-import styles from "./AboutUs.module.scss";
-
-const AboutUs = (props) => {
-  return (
-    <>
-        
-    </>
-  );
-};
-
-export default AboutUs;
