--- src/pages/ProfilePage/ProfileSubPage/AboutUs/AboutUs.module.scss
@@ -1,12 +1,0 @@
-@import "../../../../assets/main/main";
-.circle {
-    position: static;
-    background-color: #DFE2EA;
-    margin: 20px auto;
-    width: 150px;
-    height: 150px;
-    border-radius: 100px;
-  }
-.company{
-    color: getColor(primary);
-}
